+++ test/data/test_dataset.py
@@ -59,40 +59,6 @@ class TestDataset(TorchtextTestCase):
                 self.assertEqual(example.q2, expected_examples[i][1])
                 self.assertEqual(example.label, expected_examples[i][2])
 
-    def test_json_dataset_one_key_multiple_fields(self):
-        self.write_test_ppid_dataset(data_format="json")
-
-        question_field = data.Field(sequential=True)
-        spacy_tok_question_field = data.Field(sequential=True, tokenize="spacy")
-        label_field = data.Field(sequential=False)
-        fields = {"question1": [("q1", question_field),
-                                ("q1_spacy", spacy_tok_question_field)],
-                  "question2": [("q2", question_field),
-                                ("q2_spacy", spacy_tok_question_field)],
-                  "label": ("label", label_field)}
-        dataset = data.TabularDataset(
-            path=self.test_ppid_dataset_path, format="json", fields=fields)
-        expected_examples = [
-            (["When", "do", "you", "use", "シ", "instead", "of", "し?"],
-             ["When", "do", "you", "use", "シ", "instead", "of", "し", "?"],
-             ["When", "do", "you", "use", "\"&\"",
-              "instead", "of", "\"and\"?"],
-             ["When", "do", "you", "use", "\"", "&", "\"",
-              "instead", "of", "\"", "and", "\"", "?"], "0"),
-            (["Where", "was", "Lincoln", "born?"],
-             ["Where", "was", "Lincoln", "born", "?"],
-             ["Which", "location", "was", "Abraham", "Lincoln", "born?"],
-             ["Which", "location", "was", "Abraham", "Lincoln", "born", "?"],
-             "1"),
-            (["What", "is", "2+2"], ["What", "is", "2", "+", "2"],
-             ["2+2=?"], ["2", "+", "2=", "?"], "1")]
-        for i, example in enumerate(dataset):
-            self.assertEqual(example.q1, expected_examples[i][0])
-            self.assertEqual(example.q1_spacy, expected_examples[i][1])
-            self.assertEqual(example.q2, expected_examples[i][2])
-            self.assertEqual(example.q2_spacy, expected_examples[i][3])
-            self.assertEqual(example.label, expected_examples[i][4])
-
     def test_json_valid_and_invalid_nested_key(self):
         self.write_test_nested_key_json_dataset()
         valid_fields = {'foods.vegetables.name': ('vegs', data.Field()),
@@ -199,43 +165,6 @@ class TestDataset(TorchtextTestCase):
             data_iter = data.Iterator(dataset, batch_size=1,
                                       sort_within_batch=False, repeat=False)
             next(data_iter.__iter__())
-
-    def test_csv_file_no_header_one_col_multiple_fields(self):
-        self.write_test_ppid_dataset(data_format="csv")
-
-        question_field = data.Field(sequential=True)
-        spacy_tok_question_field = data.Field(sequential=True, tokenize="spacy")
-        label_field = data.Field(sequential=False)
-        # Field name/value as nested tuples
-        fields = [("ids", None),
-                  (("q1", "q1_spacy"), (question_field, spacy_tok_question_field)),
-                  (("q2", "q2_spacy"), (question_field, spacy_tok_question_field)),
-                  ("label", label_field)]
-        dataset = data.TabularDataset(
-            path=self.test_ppid_dataset_path, format="csv", fields=fields)
-        expected_examples = [
-            (["When", "do", "you", "use", "シ", "instead", "of", "し?"],
-             ["When", "do", "you", "use", "シ", "instead", "of", "し", "?"],
-             ["When", "do", "you", "use", "\"&\"",
-              "instead", "of", "\"and\"?"],
-             ["When", "do", "you", "use", "\"", "&", "\"",
-              "instead", "of", "\"", "and", "\"", "?"], "0"),
-            (["Where", "was", "Lincoln", "born?"],
-             ["Where", "was", "Lincoln", "born", "?"],
-             ["Which", "location", "was", "Abraham", "Lincoln", "born?"],
-             ["Which", "location", "was", "Abraham", "Lincoln", "born", "?"],
-             "1"),
-            (["What", "is", "2+2"], ["What", "is", "2", "+", "2"],
-             ["2+2=?"], ["2", "+", "2=", "?"], "1")]
-        for i, example in enumerate(dataset):
-            self.assertEqual(example.q1, expected_examples[i][0])
-            self.assertEqual(example.q1_spacy, expected_examples[i][1])
-            self.assertEqual(example.q2, expected_examples[i][2])
-            self.assertEqual(example.q2_spacy, expected_examples[i][3])
-            self.assertEqual(example.label, expected_examples[i][4])
-
-        # 6 Fields including None for ids
-        assert len(dataset.fields) == 6
 
     @unittest.skipIf(sys.platform == "win32", "FIXME: tempfile could not be opened twice on Windows")
     def test_csv_dataset_quotechar(self):
