+++ test/test_utils.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+# Note that all the tests in this module require dataset (either network access or cached)
 import os
 from torchtext import utils
 from .common.torchtext_test_case import TorchtextTestCase
